Added DTO and Auto Mapper
A data Transfer Object is usually an instance of a POCO (plain old CLR
object) class used as a container to encapsulate data and pass it from
one layer of the application to another.

The AutoMapper in C# is a mapper between two objects. That is AutoMapper
is an object-object mapper. It maps the properties of two different
objects by transforming the input object of one type to the output
object of another type.